Feature: Toggle interaction limitations on coupons and discounts
This feature allows customization of the interaction between coupons and discounts that are applied to the cart of the user. Discounts can disallow usage of coupons and vice versa.

Also added in this feature is a fix for the validation exception response in the cart, this would show the error message related to "code" to both coupons add input field and the gift card add input field.

Closes: feature/coupon-and-discount-interaction-limitations